refactor(chat): extract threadId-driven ChatThreadView from Conversations
Pull the transcript scroll body (message list, reactions, attachments,
past-turn insights, tool timeline / agent-insights anchor, streaming +
interrupted + parallel previews, inference status) and the three
transcript-driven modals (BackgroundProcessesPanel, SubagentDrawer,
AgentProcessSourcePanel) out of the 3451-line Conversations monolith into
a reusable, threadId-prop-driven <ChatThreadView>.

Conversations renders it for selectedThreadId; the home chat is
byte-identical (same testids/classes/behavior). This unblocks reusing the
full rich transcript on a second surface (the workflow copilot) driven by
its own dedicated thread rather than the global selectedThreadId.

Verified: pnpm typecheck clean; 120 Conversations tests + 4 ChatThreadView
tests pass.
